% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\docType{data}
\name{dictionary}
\alias{dictionary}
\title{Data dictionary}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 6 columns.
}
\usage{
dictionary
}
\description{
Data dictionary
}
\examples{
dictionary
}
\keyword{datasets}
