% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.database-package.R
\docType{package}
\name{ds.database-package}
\alias{ds.database}
\alias{ds.database-package}
\title{ds.database: Data Science Incubator Series About Working With Relational Data}
\description{
Data science incubator series about working with relational data.
}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com}

}
\keyword{internal}
