% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dm.R
\name{read_dm}
\alias{read_dm}
\title{Read the data model}
\usage{
read_dm(dir)
}
\arguments{
\item{dir}{Path to the directory containing the .csv tables that make up the
database.}
}
\value{
An object of class "dm".
}
\description{
Read the data model
}
\examples{
\dontrun{
read_dm("/path/to/csv/tables")
}
}
